Rework to python3 and using pefile
The hashbuild.py script is updated to python 3.9 (can be adjusted to
support older python 3 versions). Further, instead of parsing the
PE file with some hardcoded offsets, the pefile package is used to do
the job.
However, when parsing the relocation section, pefile is much
slower and takes a huge amount of memory for large files. It was tried
to extract the relevant components from the pefile source code, but only
minor improvements. For this reason, the original implementation is
utilized for the parsing of the relocation section (it can be switched
easily by commenting the corresponding parts in/out).